+++ src/main/java/imdl/eclesia/controller/MusicaController.java
@@ -10,6 +10,9 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 import java.util.UUID;
 
+import imdl.eclesia.domain.exception.UnauthorizedException;
+import imdl.eclesia.service.AppService;
+
 @CrossOrigin
 @RestController
 @RequestMapping("v1/musicas")
@@ -17,9 +20,11 @@ import java.util.UUID;
 public class MusicaController {
 
     private final MusicaService musicaService;
+    private final AppService appService;
 
-    public MusicaController(MusicaService musicaService) {
+    public MusicaController(MusicaService musicaService, AppService appService) {
         this.musicaService = musicaService;
+        this.appService = appService;
     }
 
     @GetMapping
@@ -37,18 +42,21 @@ public class MusicaController {
     @PostMapping
     @Operation(summary = "Adiciona uma nova música ao banco.")
     public Musica addMusica(@RequestBody MusicaInput input){
+        if (!appService.isAdminOrLider()) throw new UnauthorizedException("Apenas administradores e líderes podem adicionar músicas.");
         return musicaService.addMusica(input);
     }
 
     @PutMapping("/{id}")
     @Operation(summary = "Atualiza uma música.")
     public Musica updateMusica(@PathVariable UUID id, @RequestBody MusicaInput input){
+        if (!appService.isAdminOrLider()) throw new UnauthorizedException("Apenas administradores e líderes podem editar músicas.");
         return musicaService.updateMusica(id, input);
     }
 
     @DeleteMapping("/{id}")
     @Operation(summary = "Deletar uma música.")
     public void deleteMusica(@PathVariable UUID id){
+        if (!appService.isAdminOrLider()) throw new UnauthorizedException("Apenas administradores e líderes podem deletar músicas.");
         musicaService.deleteMusica(id);
     }
